% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotekNeo2Functions.R
\name{biotek_parse}
\alias{biotek_parse}
\title{Parser for Biotek Neo 2 plate reader data}
\usage{
biotek_parse(data_xl, layout_csv, timeseries = T)
}
\arguments{
\item{data_xl}{path to xls or xlsx file from Biotek Neo 2 plate reader}

\item{layout_csv}{path to csv file containing plate layout information}

\item{timeseries}{Boolean flag indicating whether the data is a timeseries or
single recording. Currently only works for timeseries=T.}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parser for Biotek Neo 2 plate reader data
}
\examples{
\dontrun{
biotek_parse(data_xl = "examples/plate_reader/biotek_neo2/biotek-neo2_example_data.xlsx",
             layout_csv = "examples/plate_reader/biotek_neo2/layout.csv",
             timeseries = TRUE)
}
}
