% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{calibrate_od}
\alias{calibrate_od}
\title{Convert arbitrary absorbance units to calibrated units}
\usage{
calibrate_od(pr_data, od_name, conversion_factors_csv)
}
\arguments{
\item{pr_data}{a data.frame of parsed plate reader data}

\item{od_name}{the column name for the optical density data}

\item{conversion_factors_csv}{if to_MEFL=T, path of the csv file containing
conversion factors from plate reader calibration}
}
\description{
Convert arbitrary absorbance units to calibrated units
}
