% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{calibrate_plate}
\alias{calibrate_plate}
\title{Convert arbitrary fluorescence units to calibrated units}
\usage{
calibrate_plate(pr_data, lid, gfp_gain, od_name, conversion_factors_csv)
}
\arguments{
\item{pr_data}{a data.frame of parsed plate reader data}

\item{lid}{if to_MEFL=T, the sealing mechanism used on the microtitre-plate.
See the conversion_factors_csv for available options}

\item{gfp_gain}{if to_MEFL=T, the gain value at which GFP was recorded}

\item{od_name}{the column name for the optical density data}

\item{conversion_factors_csv}{if to_MEFL=T, path of the csv file containing
conversion factors from plate reader calibration}
}
\value{

}
\description{
Convert arbitrary fluorescence units to calibrated units
}
