% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotekCytationFunctions.R
\name{cytation_parse}
\alias{cytation_parse}
\title{Parser for Biotek Cytation plate reader data}
\usage{
cytation_parse(data_csv, layout_csv, timeseries = T)
}
\arguments{
\item{data_csv}{path to a .csv, .xls or .xlsx file from Biotek Cytation plate reader}

\item{layout_csv}{path to csv file containing plate layout information}

\item{timeseries}{Boolean flag indicating whether the data is a timeseries or
single recording.}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parser for Biotek Cytation plate reader data
}
\examples{
\dontrun{
cytation_parse(data_csv = "examples/plate_reader/biotek_cytation/231031_pr_calibration_lid.xlsx",
                layout_csv = "examples/plate_reader/biotek_cytation/plate_layout.csv",
                timeseries = FALSE)
}
}
