% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCoreMisc.R
\name{ff_append_cols}
\alias{ff_append_cols}
\title{Append data columns to a flowFrame}
\usage{
ff_append_cols(fr, cols)
}
\arguments{
\item{fr}{A \code{flowFrame} (see \pkg{flowCore}).}

\item{cols}{A numeric matrix containing the new data columns to be added.
Must has column names to be used as new channel names.}
}
\value{
A \code{flowFrame} (see \pkg{flowCore})
}
\description{
Append data columns to a flowFrame
}
\details{
It is used to add extra data columns to the existing flowFrame.  It handles
keywords and parameters properly to ensure the new flowFrame can be written
as a valid FCS through the function \code{write.FCS} .
}
\author{
Mike Jiang
}
\keyword{IO}
