% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{generateCfs}
\alias{generateCfs}
\title{Generate conversion factors}
\usage{
generateCfs(calibration_dir, date)
}
\arguments{
\item{calibration_dir}{path of the directory in which your calibration data
is held}

\item{date}{input date as YYYYMMDD on which calibration was carried out.
n.b. this should also be the name of the folder holding the csv files}
}
\value{
saves a csv data frame with the conversion factors and two png plots
of the conversion factors (for Absorbance and fluorescence)
}
\description{
Generate conversion factors
}
