% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{get_bacteria}
\alias{get_bacteria}
\title{Remove background debris events}
\usage{
get_bacteria(flow_frame, pre_cleaned)
}
\arguments{
\item{flow_frame}{a \code{flowFrame} for processing}

\item{pre_cleaned}{a flag to identify if debris has already been manually
gated}
}
\description{
Remove background debris events
}
