% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{get_calibration}
\alias{get_calibration}
\title{Get fluorescence calibration curve parameters}
\usage{
get_calibration(bead_file, flu_channels, mef_peaks, manual_peaks, bead_dens_bw)
}
\arguments{
\item{bead_file}{path to beads .fcs file.}

\item{flu_channels}{a vector of strings of the fluorescence channels to keep
in the trimmed data and plotting. Defaults to \code{"BL1-H"}.}

\item{mef_peaks}{a list of lists in the form \code{list(list(channel="BL1-H",
peaks=c(0, 200, ...)} of MEF fluorescence values for the calibration beads.
 Default values for BL1-H and YL2-H.}

\item{manual_peaks}{if bead peaks are not being found by the EM algorithm,
one can manually specify the positions of the bead peaks (on a log10 scale).
 A list of lists in the form
 \code{list(list(channel="BL1-H", peaks=c(2.1, 2.8, ...)} of log10
 fluorescence values for the calibration beads.}

\item{bead_dens_bw}{the bandwidth for the kernel density of the bead peak
data. Default = 0.025. Increase if erroneous peaks are being found, decrease
 if not enough peaks are found.}
}
\value{

}
\description{
Get fluorescence calibration curve parameters
}
