% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{get_singlets}
\alias{get_singlets}
\title{Remove doublet events}
\usage{
get_singlets(flow_frame)
}
\arguments{
\item{flow_frame}{a \code{flowFrame} for processing}
}
\description{
Remove doublet events
}
