% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecanInfiniteFunctions.R,
%   R/tecanSparkFunctions.R
\name{next_blank}
\alias{next_blank}
\title{Find next blank line}
\usage{
next_blank(start_idx, data)

next_blank(start_idx, data)
}
\arguments{
\item{data}{}
}
\value{
row index of next blank line

row index of next blank line
}
\description{
Find next blank line

Find next blank line
}
