% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecanSparkFunctions.R
\name{next_blank_row}
\alias{next_blank_row}
\title{Find next blank line}
\usage{
next_blank_row(start_idx, data)
}
\arguments{
\item{data}{}
}
\value{
row index of next blank line
}
\description{
Find next blank line
}
