% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecanInfiniteFunctions.R,
%   R/tecanSparkFunctions.R
\name{next_filled}
\alias{next_filled}
\title{Find next non-blank line}
\usage{
next_filled(start_idx, data)

next_filled(start_idx, data)
}
\arguments{
\item{data}{}
}
\value{
row index of next non-blank line

row index of next non-blank line
}
\description{
Find next non-blank line

Find next non-blank line
}
