% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecanSparkFunctions.R
\name{next_filled_cell}
\alias{next_filled_cell}
\title{Find next non-blank line}
\usage{
next_filled_cell(start_idx, data, col)
}
\arguments{
\item{data}{}
}
\value{
row index of next non-blank line
}
\description{
Find next non-blank line
}
