% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{od_norm}
\alias{od_norm}
\title{Normalisation absorbance against blank well}
\usage{
od_norm(pr_data, blank_well, od_name)
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data}

\item{blank_well}{the well coordinates of one or more media blanks}

\item{od_name}{the column name for the optical density data}
}
\value{
an updated data.frame with an additional column "normalised_OD"
}
\description{
Normalisation absorbance against blank well
}
