% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{prep_flow_frame}
\alias{prep_flow_frame}
\title{Log10 transform and remove -Inf values from \code{flowFrame}}
\usage{
prep_flow_frame(flow_frame, flu_channels)
}
\arguments{
\item{flow_frame}{a \code{flowFrame} for processing}

\item{flu_channels}{a vector of strings of the fluorescence channels to keep
in the trimmed data and plotting. Defaults to \code{"BL1-H"}.}
}
\value{

}
\description{
Log10 transform and remove -Inf values from \code{flowFrame}
}
