% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{prNorm}
\alias{prNorm}
\title{Plate reader normalisation and calibration}
\usage{
prNorm(
  pr_data,
  blank_well = "A1",
  neg_well = c("A2"),
  OD_name = "OD",
  flu_names = c("GFP"),
  af_model = "polynomial",
  to_MEFL = F,
  GFP_gain = 0,
  lid_type = "nolid",
  conversion_factors_csv = NA
)
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data}

\item{blank_well}{the well coordinates of one or more media blanks}

\item{neg_well}{the well coordinates of a non-fluorescent control}

\item{OD_name}{the column name for the optical density data}

\item{flu_names}{the column names for the fluorscence data}

\item{af_model}{model used to fit negative control autofluorescence.
For now these include "polynomial", "invers_poly", "exponential", "spline" or "loess".}

\item{to_MEFL}{a Boolean to determine whether to attempt to convert OD and
GFP reading to calibrated units}

\item{GFP_gain}{if to_MEFL=T, the gain value at which GFP was recorded}

\item{lid_type}{if to_MEFL=T, the sealing mechanism used on the
microtitre-plate. See the conversion_factors_csv for available options}

\item{conversion_factors_csv}{if to_MEFL=T, path of the csv file containing
conversion factors from plate reader calibration}
}
\value{
a data.frame with columns for raw plate reader data, normalised data
and, if to_MEFL = T, calibrated OD and GFP data
}
\description{
Deprecated. Please use \code{process_plate()}.
}
