% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{summarise_data}
\alias{summarise_data}
\title{Get geometric mean and standard deviation of fluorescence channels of a flow frame}
\usage{
summarise_data(flow_frame, flu_channels)
}
\arguments{
\item{flow_frame}{a flow frame to be summarised}

\item{flu_channels}{channels to summarise}
}
\value{

}
\description{
Get geometric mean and standard deviation of fluorescence channels of a flow frame
}
