% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{trim.fcs}
\alias{trim.fcs}
\title{Process .fcs files within a directory to remove debris and doublets, and
optionally calibrate}
\usage{
trim.fcs(
  dir_path,
  pattern = "*.fcs",
  flu_channels = c("BL1-H"),
  do_plot = F,
  pre_cleaned = F,
  calibrate = F,
  MEF_peaks = NA,
  bead_dens_bw = 0.025,
  manual_peaks = NA
)
}
\arguments{
\item{dir_path}{a directory path containing the .fcs files  to be parsed or
folders to be recursed through.}

\item{pattern}{a regex pattern to match particular .fcs files. Default is
\code{"*.fcs"} matching all .fcs files.}

\item{flu_channels}{a list of strings of the fluorescence channels to keep
in the trimmed data and plotting. Defaults to "BL1-H".}

\item{do_plot}{a Boolean flag to determine whether to produce plots showing
the trimming of each \code{flowFrame}. Defaults to \code{FALSE}.}

\item{pre_cleaned}{have you pre removed background debris}

\item{calibrate}{a Boolean flag to determine whether to convert fluorescence
to MEF values. Requires an .fcs file with named \code{"*beads*.fcs"}.
Defaults to \code{FALSE}.}

\item{MEF_peaks}{a list of lists in the form \code{list(list(channel="BL1-H",
peaks=c(0, 200, ...)} of MEF fluorescence values for the calibration beads.
 Default values for BL1-H and YL2-H.}

\item{bead_dens_bw}{the bandwidth for the kernel density of the bead peak
data. Default = 0.025. Increase if erroneous peaks are being found, decrease
 if not enough peaks are found.}

\item{manual_peaks}{if bead peaks are not being found by the EM algorithm,
one can manually specify the positions of the bead peaks (on a log10 scale).
 A list of lists in the form
 \code{list(list(channel="BL1-H", peaks=c(2.1, 2.8, ...)} of log10
 fluorescence values for the calibration beads.}
}
\value{
nothing is returned. A new folder is created with the trimmed .fcs
files and plot if \code{do_plot = TRUE}.
}
\description{
Deprecated, please use \code{process_fcs_dir}. This function will remain for
backwards compatibility.
}
\details{
\code{trim.fcs} uses mixture models to cluster bacteria from background
debris and fits a linear model to SSC-H vs SSC-A to remove doublets.
}
