% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotekCytationFunctions.R
\name{cytation_parse}
\alias{cytation_parse}
\title{Parser for Biotek Cytation plate reader data}
\usage{
cytation_parse(data_csv, layout_csv, timeseries = T)
}
\arguments{
\item{layout_csv}{path to csv file containing plate layout information}

\item{timeseries}{Boolean flag indicating whether the data is a timeseries or
single recording. Currently only works for timeseries=T.}

\item{data_xl}{path to xls or xlsx file from Biotek Cytation plate reader}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parser for Biotek Cytation plate reader data
}
