% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{flu_norm}
\alias{flu_norm}
\title{Normalise fluorescence against negative well}
\usage{
flu_norm(pr_data, neg_well, blank_well, flu_name, af_model, data_csv)
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data with OD
normalised}

\item{neg_well}{the well coordinates of a non-fluorescent control}

\item{blank_well}{the well coordinates of a media blank}

\item{flu_name}{the column name of the fluorescence chanel to normalise}

\item{af_model}{model used to fit negative control autofluorescence.
For now these include "polynomial", "invers_poly", "exponential", "spline" or "loess".}

\item{data_csv}{path to the original data. Used for saving normalisation curve plots.}
}
\value{

}
\description{
Normalise fluorescence against negative well
}
