% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{generate_cfs}
\alias{generate_cfs}
\title{Generate Conversion Factors}
\usage{
generate_cfs(calibration_csv)
}
\arguments{
\item{calibration_csv}{path of a .csv file of your calibration data}
}
\description{
Generate Conversion Factors
}
