% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecanInfiniteFunctions.R
\name{infinite_parse}
\alias{infinite_parse}
\title{Parser for Tecan Infinite plate reader data}
\usage{
infinite_parse(data_csv, layout_csv, timeseries = F)
}
\arguments{
\item{data_csv}{path to .csv, .xls or .xlsx file from Tecan Infinit plate reader}

\item{layout_csv}{path to csv file containing plate layout information}

\item{timeseries}{Boolean flag indicating whether the data is a timeseries or
single recording. The Tecan software outputs the two scenarios differently.}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parser for Tecan Infinite plate reader data
}
\examples{
\dontrun{
infinite_parse(data_csv = "examples/plate_reader/tecan_infinite/calibration_data.csv",
                layout_csv = "examples/plate_reader/tecan_infinite/calibration_plate_layout.csv",
                timeseries = TRUE)
}
}
