% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{plot_trimming}
\alias{plot_trimming}
\title{Plot trimming process}
\usage{
plot_trimming(
  flow_frame,
  bacteria_flow_frame,
  singlet_flow_frame,
  normalised_flow_frame,
  calibrated_flow_frame,
  flu_channels,
  out_name,
  normalised,
  calibrate
)
}
\arguments{
\item{flow_frame}{the original \code{flowFrame}}

\item{bacteria_flow_frame}{the \code{flowFrame} with debris removed}

\item{singlet_flow_frame}{the \code{flowFrame} with debris and doublets removed}

\item{normalised_flow_frame}{the normalised \code{flowFrame}}

\item{calibrated_flow_frame}{the calibrated \code{flowFrame} with debris and doublets
removed}

\item{flu_channels}{a vector of strings of the fluorescence channels to keep
in the processed data and plotting. Defaults to "BL1-H".}

\item{out_name}{the filename for the outputted \code{flowFrame}}

\item{normalised}{a Boolean flag to determine whether to normalise}

\item{calibrate}{a Boolean flag to determine whether to convert fluorescence
to MEF values. Requires an .fcs file with named \code{"*beads*.fcs"}.
Defaults to \code{FALSE}.}
}
\description{
Plot trimming process
}
