% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCytometryFunctions.R
\name{process_fcs}
\alias{process_fcs}
\title{Process an .fcs files to remove debris and doublets}
\usage{
process_fcs(fcs_file, flu_channels = c("BL1-H"), do_plot = F, pre_cleaned = F)
}
\arguments{
\item{fcs_file}{path of an .fcs files to be processed.}

\item{flu_channels}{a vector of strings of the fluorescence channels to keep
in the trimmed data and plotting. Defaults to "BL1-H".}

\item{do_plot}{a Boolean flag to determine whether to produce plots showing
the trimming of each \code{flowFrame} Defaults to \code{FALSE}.}

\item{pre_cleaned}{have you pre removed background debris.}
}
\value{
nothing is returned. A trimmed .fcs file is produced and a plot if
\code{do_plot = TRUE}.
}
\description{
\code{process_fcs} uses mixture models to cluster bacteria from background
debris and fits a linear model to SSC-H vs SSC-A to remove doublets.
}
