% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReaderFunctions.R
\name{process_plate}
\alias{process_plate}
\title{Plate reader normalisation and calibration}
\usage{
process_plate(
  data_csv,
  blank_well = "A1",
  neg_well = "A2",
  od_name = "OD",
  flu_names = c("GFP"),
  af_model = "polynomial",
  to_MEFL = F,
  flu_gains,
  conversion_factors_csv
)
}
\arguments{
\item{data_csv}{path to a .csv file containing parsed plate reader data}

\item{blank_well}{the well coordinates of one or more media blanks}

\item{neg_well}{the well coordinates of a non-fluorescent control}

\item{od_name}{the column name for the optical density data}

\item{flu_names}{the column names for the fluorescence data}

\item{af_model}{model used to fit negative control autofluorescence.
For now these include "polynomial", "invers_poly", "exponential", "spline" and "loess".}

\item{to_MEFL}{a Boolean to determine whether to attempt to convert OD and
GFP reading to calibrated units}

\item{flu_gains}{if to_MEFL=T, the gain values at which the fluorophores
specified in flu_names was recorded. If there isn't calibration data for a
fluorophore, do not speficy a gain value}

\item{conversion_factors_csv}{if to_MEFL=T, path of the csv file containing
conversion factors from plate reader calibration}
}
\value{
a data.frame with columns for raw plate reader data, normalised data
and, if to_MEFL = T, calibrated OD and GFP data
}
\description{
Plate reader normalisation and calibration
}
