% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{sparkParse}
\alias{sparkParse}
\title{Parser for Tecan Spark plate reader data}
\usage{
sparkParse(data_csv, layout_csv)
}
\arguments{
\item{data_csv}{path to csv file from Tecan Spark plate reader}

\item{layout_csv}{path to csv file containing plate layout information}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parser for Tecan Spark plate reader data
}
